requirements.txt
python-dotenv==1.0.1
python_fasthtml==0.6.10
pytz==2023.3
pytz==2024.2
supabase==2.9.0
